# ToSlicePtr: reduce allocations and improving speed