Hi, I found a way to reduce the number of allocations and improve the overall performance of `ToSlicePtr`.

here is simple benchmark:
```
name           old time/op    new time/op    delta
ToSlicePtr-10    1.18ms ± 6%    0.25ms ± 5%   -78.60%  (p=0.000 n=14+12)

name           old alloc/op   new alloc/op   delta
ToSlicePtr-10    1.60MB ± 0%    0.80MB ± 0%   -49.92%  (p=0.000 n=15+15)

name           old allocs/op  new allocs/op  delta
ToSlicePtr-10      100k ± 0%        0k ± 0%  -100.00%  (p=0.000 n=15+15)
```