@andig 2024-06-29
Out of curiosity: why does this have an effect? `Map` already pre-allocates the resulting slice- why isn't this sufficient?